requirements.txt
django
djangorestframework
mysqlclient
pillow
djangorestframework-simplejwt
drf-yasg
django-widget-tweaks
django-import-export